+++ tests/tests_dfa_haystack_and_needle.py
@@ -1,5 +1,5 @@
 import unittest
-from src.DFAHaystackAndNeedle import search_num_of_in
+from src.dfa_haystack_and_needle import search_num_of_in
 
 
 class TestSearchNumOfIn(unittest.TestCase):
