+++ tests/tests_hamsters_shops.py
